Rename filenames to "signed-data"
