+++ packages/airnode-deployer/terraform/airnode/aws/templates/signedDataApiGateway.yaml.tpl
