+++ packages/airnode-deployer/terraform/airnode/gcp/templates/signedDataApiGateway.yaml.tpl
